requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.6.0
django-mysql==3.10.0
djangorestframework==3.12.2
PyMySQL==1.0.2
pytz==2020.5
PyYAML==5.3.1
sqlparse==0.4.1
requests==2.25.1
tqdm==4.56.0
xgboost==1.4.0
numpy==1.19.5
scipy==1.5.4
